@@ -3,7 +3,11 @@ package com.hstudio.jlog
 import android.util.Log
 
 object JLog {
-    private val IS_DEBUG = BuildConfig.DEBUG
+    private var isPrintLog = true;
+
+    fun setLogPrint(isPrintLog: Boolean) {
+        this.isPrintLog = isPrintLog
+    }
 
     @JvmStatic
     fun printStackTrace() {
@@ -18,36 +22,71 @@ object JLog {
     }
 
     @JvmStatic
-    fun d(tag: String, message: String? = null) {
-        if (!IS_DEBUG) return
+    fun d(message: String) {
+        if (!isPrintLog) return
+        val log = beautifyLog(message)
+        Log.d("JLog", log)
+    }
+
+    @JvmStatic
+    fun d(tag: String, message: String?) {
+        if (!isPrintLog) return
         val log = beautifyLog(message)
         Log.d(tag, log)
     }
 
     @JvmStatic
-    fun e(tag: String, message: String? = null) {
-        if (!IS_DEBUG) return
+    fun e(message: String) {
+        if (!isPrintLog) return
+        val log = beautifyLog(message)
+        Log.e("JLog", log)
+    }
+
+    @JvmStatic
+    fun e(tag: String, message: String?) {
+        if (!isPrintLog) return
         val log = beautifyLog(message)
         Log.e(tag, log)
     }
 
     @JvmStatic
-    fun i(tag: String, message: String? = null) {
-        if (!IS_DEBUG) return
+    fun i(message: String) {
+        if (!isPrintLog) return
+        val log = beautifyLog(message)
+        Log.i("JLog", log)
+    }
+
+    @JvmStatic
+    fun i(tag: String, message: String?) {
+        if (!isPrintLog) return
         val log = beautifyLog(message)
         Log.i(tag, log)
     }
 
     @JvmStatic
-    fun w(tag: String, message: String? = null) {
-        if (!IS_DEBUG) return
+    fun w(message: String) {
+        if (!isPrintLog) return
+        val log = beautifyLog(message)
+        Log.w("JLog", log)
+    }
+
+    @JvmStatic
+    fun w(tag: String, message: String?) {
+        if (!isPrintLog) return
         val log = beautifyLog(message)
         Log.w(tag, log)
     }
 
     @JvmStatic
-    fun v(tag: String, message: String? = null) {
-        if (!IS_DEBUG) return
+    fun v(message: String) {
+        if (!isPrintLog) return
+        val log = beautifyLog(message)
+        Log.v("JLog", log)
+    }
+
+    @JvmStatic
+    fun v(tag: String, message: String?) {
+        if (!isPrintLog) return
         val log = beautifyLog(message)
         Log.v(tag, log)
     }
